% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biodiv_functions.R
\name{Div}
\alias{Div}
\title{Div(data)  Data.frame for value of diversity required to plot the results}
\usage{
Div(data)
}
\arguments{
\item{data}{This parameter must necessarily be an object containing the outcome of the FDchao function; the div function will filter these outcomes to prepare the graph.}
}
\value{
data.frame for value of taxonomic diversity, functional diversity and functional redundance.
}
\description{
This function prepares a "data.frame" object where it takes 5 variables, namely
(qEix, TauMin, TauMed, RedFunChao, Color).
These variables are required to plot the graphs (using the PlotDiv function)
depicting taxonomic, functional, and functional redundancy diversity results.
}
