% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biodiv_functions.R
\name{FDsingle}
\alias{FDsingle}
\title{FDsingle(data, distance, tau, q) Functional Diversity of a single site for specified values of tau and q of Hill}
\usage{
FDsingle(data, distance, tau, q)
}
\arguments{
\item{data}{a vector of species sample frequencies.}

\item{distance}{a matrix of species-pairwise distances.}

\item{tau}{a numeric for a specified level of threshold distinctiveness.}

\item{q}{a numeric for a specified diversity order q.}
}
\value{
a numeric value of Functional Diversity.
}
\description{
This function calculates functional diversity for a single sample
at a specific value of tau and Hill's q.
}
