% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biodiv_functions.R
\name{PlotDiv}
\alias{PlotDiv}
\title{PlotDiv(data) function to plot the value of diversity}
\usage{
PlotDiv(data)
}
\arguments{
\item{data}{This parameter must necessarily be an object containing the outcome of the Div function to prepare the graph.}
}
\value{
Three graphs, namely (functional diversity, taxonomic diversity, and functional redundancy).
}
\description{
This is a function to plot the results of the FDchao function
filtered by the Div function, generating three graphs:
Taxonomic Diversity, Functional Diversity, and Functional Redundancy.
}
