% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biodiv_functions.R
\name{FDchao}
\alias{FDchao}
\title{FDchao(data, distance, tau, q, boot, datatype) Functional Diversity of N sites for various values of tau and q}
\usage{
FDchao(data, distance, tau, q, boot)
}
\arguments{
\item{data}{a list with N sites; each element of list is species abundances.}

\item{distance}{a matrix of species-pairwise distances.}

\item{tau}{a numeric or a vector of levels of threshold distinctiveness.}

\item{q}{a numeric or a vector of diversity orders; the suggested range for q is \link{0, 2}.}

\item{boot}{a numeric of number of bootstrap replications.}
}
\value{
two matrices of FD; the first matrix is the q-profile information, the second matrix is the tau profile information.
}
\description{
This function calculate Functional Diversity of N sites
for various values of tau and q
}
