% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biodiv_functions.R
\name{plotDiv}
\alias{plotDiv}
\title{plotDiv(data, tog, cap) Function to plot the value of diversity}
\usage{
plotDiv(data, tog = TRUE, cap = FALSE)
}
\arguments{
\item{data}{This parameter must necessarily be an object containing the outcome of the Div function to prepare the graph.}

\item{tog}{This parameter is an abbreviation of the word "together" and is a logical object. If it is set to TRUE, the function will plot the three graphs together (side by side).}

\item{cap}{This parameter is an abbreviation of the word "captions" and is a logical object. If it is set to TRUE, the function will ask you to set the labels for the graph.}
}
\value{
Three graphs, namely (functional diversity, taxonomic diversity, and functional redundancy).
}
\description{
This is a function to plot the results of the FDchao function
filtered by the Div function, generating three graphs:
Taxonomic Diversity, Functional Diversity, and Functional Redundancy.
}
\examples{
# example code. help("plotDiv") for more details.
plotDiv(datafilt) #To use the plot, it is necessary to fulfill two prerequisites: the first one is to create an object using the initial function FDchao, the second one is to create an object with the function's argument Div being the result of step 1. After that, you can proceed to the final step, which is the plot of the graph plotDiv"
plotDiv(FiltBD, tog = TRUE, cap = TRUE) #set the labels for the graph
plotDiv(FiltBD, tog = FALSE, cap = FALSE) #Choose a plot to create
}
